update post api and user api

+++ Social/src/user/user.service.ts
@@ -1,6 +1,6 @@
 import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
-import fs from 'fs';
+const fs = require('fs')
 import {
   ForgotPasswordDTO,
   UpdatePasswordDTO,
@@ -92,12 +92,16 @@ export class UserService {
     // const url1 = 'http://localhost:8888/public/img/' + imgName;
     const url2 = 'http://camenryder.xyz/public/img/' + imgName;
     if (user.url_avatar != null) {
-      const filePath = user.url_avatar.replace(substringToRemove, '');
-      fs.unlink(filePath, (err) => {
-        if (err) {
-          throw new ForbiddenException('Old file avatar couldnt delete');
-        }
-      });
+      if (
+        user.url_avatar != 'http://camenryder.xyz/public/img/avatar_default.png'
+      ) {
+        const filePath = user.url_avatar.replace(substringToRemove, '');
+        fs.unlink(filePath, (err) => {
+          if (err) {
+            throw new ForbiddenException('Old file avatar couldnt delete');
+          }
+        });
+      }
     }
     user.url_avatar = url2;
     await this.prismaService.user.update({
